added missing cast at call to regsave, in ms_gto case

+++ mach/proto/cg/regvar.c
@@ -103,7 +103,7 @@ fixregvars(saveall) {
 		if (saveall) {
 			struct reginfo *rp;
 			rp= &machregs[rvnumbers[rvtyp][i]];
-			regsave(codestrings[rp->r_repr],-EM_WSIZE,rp->r_size);
+			regsave(codestrings[rp->r_repr],(long)-EM_WSIZE,rp->r_size);
 		} else if(regassigned[rvtyp][i].ra_score>0) {
 			rv=regassigned[rvtyp][i].ra_rv;
 			rv->rv_reg=rvnumbers[rvtyp][i];
